Cleanup to extract the sort options setup
