+++ src/hooks/useFetchHotels.tsx
@@ -1,7 +1,6 @@
 import { useCallback, useContext } from "react";
 import { HotelsContext } from "../contexts/HotelsContext";
-import { DEFAULT, HIGH, ID_KEY, LOW } from "../constants/sortKeys";
-import { TSortKeys, TSortOrder } from "../types/sort";
+import { getSortOptions } from "../utils/getSortOptions";
 
 const useFetchHotels = () => {
   const { setIsLoading, setError, setHotels, setSortOptions } =
@@ -21,26 +20,7 @@ const useFetchHotels = () => {
       setHotels(res);
 
       if (res.length) {
-        const sample = res[0];
-        const numericProperties = Object.entries(sample).filter((entry) => {
-          const key = entry[0];
-          const value = entry[1];
-          return key !== ID_KEY && typeof value === "number";
-        });
-        const sortOptions = [
-          { id: DEFAULT as TSortKeys, order: "" as TSortOrder },
-        ];
-        numericProperties.forEach((option) => {
-          sortOptions.push({
-            id: option[0] as TSortKeys,
-            order: HIGH as TSortOrder,
-          });
-          sortOptions.push({
-            id: option[0] as TSortKeys,
-            order: LOW as TSortOrder,
-          });
-        });
-        setSortOptions(sortOptions);
+        setSortOptions(getSortOptions(res));
       }
     } catch (e) {
       console.error(e);
