+++ src/interfaces/SortOption.ts
@@ -1,0 +1,6 @@
+import { TSortKeys, TSortOrder } from "../types/sort";
+
+export interface SortOption {
+  id: TSortKeys;
+  order: TSortOrder;
+}
