+++ src/screens/HomeScreen/HomeScreen.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useCallback, useState, useContext } from "react";
+import React, { useEffect, useContext } from "react";
 import {
   View,
   Text,
